requirements.txt
streamlit
langchain
langchain-core
langchain-groq
langchain-community
langchain-huggingface
langchain-text-splitters
faiss-cpu
python-dotenv
sentence-transformers